Add files via upload
note: for some reason, on my machine, mst cannot be imported without error. If one wishes to see tests run error-free with the code that is in mst, simply copy and paste the code within to the test_mst file and run. Should show all three tests running error-free, or at least it did for me.